use diminish instead of rich-minority
This isn't quite as cool because you can't see the lighters that are
being hidden, but in all the time I've used rich-minority mode, I don't
know that I've done it more than once. Besides, I don't know what all of
these strings even refer to (versus symbols of minor modes, which are
more obvious). Also, this lets me group the lighter-hiding stuff with
the packages they belong to.