Reorganize init and incorporate no-littering
This migration isn't complete yet, but it looks like it will save me
some code.

To do this, I want to pull the no-littering stuff into init.el, but
that means initializing package.el so we can put it in a
use-package. Move the initialization step for package into init.el as
a stopgap, at least so long as it's out of ethan-packages.