Try to optimize fish prompt for huge repos
On my work laptop, I have to interact with gecko-git, which is
abnormally large due to reasons beyond my control. This makes git
operations slower. Because displaying a prompt requires several git
operations, it's worth some ugliness to make each prompt faster. This
is an attempt to minimize the work done to render what I actually care
about on the left prompt, which is just a heads-up if we're doing
something scary.

I'd like to back this out one day once I'm no longer working on huge
git repos.